Fix serveral issues

+++ torch/csrc/Module.cpp
@@ -341,7 +341,7 @@ static PyObject * TH_CONCAT_2(THPModule_, name)(PyObject *_unused, PyObject *arg
   PyObject *tensor = THPDefaultTensorClass;                                    \
   PyObject *key, *value;                                                       \
   Py_ssize_t pos = 0;                                                          \
-  for (int i = (int) PyTuple_Size(args)-1; i >= 0; i--) {                            \
+  for (Py_ssize_t i = PyTuple_Size(args)-1; i >= 0; i--) {                            \
     PyObject *item = PyTuple_GET_ITEM(args, i);                                \
     if (THPModule_isTensor(item) || THPVariable_Check(item)) {                 \
       tensor = item;                                                           \
@@ -496,7 +496,7 @@ PyObject *THPModule_inferSize(PyObject *_unused, PyObject *args)
   THLongStorage *sizes = sizes_guard.get();
 
   char error_buffer[1024];
-  int ret = (int) THLongStorage_inferSize2(sizes, size1->data, size1->size, size2->data, size2->size, error_buffer, 1024);
+  int ret = THLongStorage_inferSize2(sizes, size1->data, size1->size, size2->data, size2->size, error_buffer, 1024);
   THPUtils_assert(ret == 0, error_buffer);
   return THPSize_New(sizes->size, sizes->data);
   END_HANDLE_TH_ERRORS
